Moved test_data into test_config.
