+++ tests/test_config.py
@@ -1,16 +1,15 @@
+import re
 from pathlib import Path
-from pydantic import ValidationError
+
 import pytest
+from pydantic import ValidationError
 
+from .. import copier
 from ..copier.config.factory import make_config
-from ..copier.config.objects import (
-    ConfigData,
-    EnvOps,
-    Flags,
-    DEFAULT_DATA,
-    DEFAULT_EXCLUDE,
-)
-
+from ..copier.config.objects import (DEFAULT_DATA, DEFAULT_EXCLUDE, ConfigData,
+                                     EnvOps, Flags)
+from ..copier.config.user_data import (load_config_data, load_json_data,
+                                       load_toml_data, load_yaml_data)
 
 GOOD_FLAGS = {
     "pretend": True,
@@ -29,6 +28,55 @@ GOOD_ENV_OPS = {
     "keep_trailing_newline": False,
     "i_am_not_a_member": None,
 }
+
+
+def test_config_data_is_loaded_from_file():
+    config = load_config_data("tests/demo_data")
+    config["_exclude"] = ["exclude1", "exclude2"]
+    config["_include"] = ["include1", "include2"]
+    config["_skip_if_exists"] = ["skip_if_exists1", "skip_if_exists2"]
+    config["_tasks"] = ["touch 1", "touch 2"]
+    config["_extra_paths"] = ["test"]
+
+
+@pytest.mark.parametrize(
+    "template",
+    ["tests/demo_toml", "tests/demo_yaml", "tests/demo_yml", "tests/demo_json"],
+)
+def test_read_data(dst, template):
+    copier.copy(template, dst, force=True)
+
+    gen_file = dst / "user_data.txt"
+    result = gen_file.read_text()
+    expected = Path("tests/user_data.ref.txt").read_text()
+    assert result == expected
+
+
+def test_bad_toml(capsys):
+    assert {} == load_toml_data("tests/demo_badtoml")
+
+
+def test_invalid_toml(capsys):
+    assert {} == load_yaml_data("tests/demo_invalid")
+    out, _ = capsys.readouterr()
+    assert re.search(r"INVALID.*tests/demo_invalid/copier\.yml", out)
+
+    assert {} == load_toml_data("tests/demo_invalid")
+    out, _ = capsys.readouterr()
+    assert re.search(r"INVALID.*tests/demo_invalid/copier\.toml", out)
+
+    assert {} == load_json_data("tests/demo_invalid")
+    out, _ = capsys.readouterr()
+    assert re.search(r"INVALID.*tests/demo_invalid/copier\.json", out)
+
+    assert {} == load_config_data("tests/demo_invalid", _warning=False)
+    assert re.search(r"INVALID", out)
+
+
+def test_invalid_quiet(capsys):
+    assert {} == load_config_data("tests/demo_invalid", quiet=True)
+    out, _ = capsys.readouterr()
+    assert out == ""
 
 
 # Flags
